fixes #2411 refactor config/initializers
