+++ config/initializers/foreman.rb
@@ -1,4 +1,17 @@
-require 'foreman'
+require 'foreman/access_permissions'
+require 'foreman/default_data/loader'
+require 'foreman/renderer'
+require 'foreman/controller'
+require 'net'
+require 'foreman/provision' if SETTINGS[:unattended]
+require 'audit_extensions'
+
+module Foreman
+  # generate a UUID
+  def self.uuid
+    UUIDTools::UUID.random_create.to_s
+  end
+end
 
 # We may be executing something like rake db:migrate:reset, which destroys this table
 # only continue if the table exists
