--- config/initializers/hostable.rb
@@ -1,3 +1,0 @@
-class ActiveRecord::Base
-  extend Host::Hostmix
-end
