+++ config/initializers/wrap_parameters.rb
@@ -4,10 +4,10 @@
 
 # Enable parameter wrapping for JSON. You can disable this by setting :format to an empty array.
 ActiveSupport.on_load(:action_controller) do
- wrap_parameters :format => [:json]
+  wrap_parameters :format => [:json]
 end
 
 # Disable root element in JSON by default.
 ActiveSupport.on_load(:active_record) do
- self.include_root_in_json = true
+  self.include_root_in_json = true
 end
